#92038 Create view and viewPaneContainer util actions
